Import ccnmtlsettings and adapt some functionality
I am in the process of migrating ccnmtlsettings to this new library,
`ctlsettings`, as it just seems to make sense - there are a number of
breaking changes I want to introduce, and this new library will allow
us to push these changes out gradually. We also might as well finally
rename this as we haven't been the CCNMTL since 2015.

* Removed django-markwhat, django-compressor, djangowind,
  and their configuration settings.
* Updated a few instances of the 'ccnmtl' string to 'ctl'

Next steps: introduce a generalized django-cas-ng configuration into this
library, based on one of our existing django apps that use it, such as
econplayground.