Remove statsd, django-statsd-mozilla
So, `django-statsd-mozilla` is unmaintained, and I have seen some issues
when attempting to install this package in newer environments, with
`statsd` as well. These metrics which are being sent to graphite are
largely going unused, to my knowledge. I propose removing this for now
and re-thinking this in the future if necessary.

Our django-smoketest package doesn't rely on this, and hound *should*
continue to function as normal, but I could be wrong. We'll see.